// frontend/src/app/api/models/pagina-termino-glosario-publico.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PaginaMeta } from '../models/pagina-meta';
import { TerminoGlosarioPublico } from '../models/termino-glosario-publico';
export type PaginaTerminoGlosarioPublico = PaginaMeta & {
'resultados': Array<TerminoGlosarioPublico>;
};